+++ src/main/g8/app/forms/mappings/Mappings.scala
@@ -1,10 +1,11 @@
 package forms.mappings
 
-import java.time.LocalDate
-
+import models.Enumerable
 import play.api.data.FieldMapping
 import play.api.data.Forms.of
-import models.Enumerable
+import play.api.i18n.Messages
+
+import java.time.LocalDate
 
 trait Mappings extends Formatters with Constraints {
 
@@ -33,6 +34,6 @@ trait Mappings extends Formatters with Constraints {
                            allRequiredKey: String,
                            twoRequiredKey: String,
                            requiredKey: String,
-                           args: Seq[String] = Seq.empty): FieldMapping[LocalDate] =
+                           args: Seq[String] = Seq.empty)(implicit messages: Messages): FieldMapping[LocalDate] =
     of(new LocalDateFormatter(invalidKey, allRequiredKey, twoRequiredKey, requiredKey, args))
 }