+++ src/main/g8/app/viewmodels/govuk/DateFluency.scala
@@ -27,7 +27,17 @@ trait DateFluency {
                fieldset: Fieldset
              )(implicit messages: Messages): DateInput = {
 
-      val errorClass = if (errorMessage(field).isDefined) "govuk-input--error" else ""
+      val errorClass = "govuk-input--error"
+
+      val dayError         = field.error.exists(_.args.contains(messages("date.error.day")))
+      val monthError       = field.error.exists(_.args.contains(messages("date.error.month")))
+      val yearError        = field.error.exists(_.args.contains(messages("date.error.year")))
+      val anySpecificError = dayError || monthError || yearError
+      val allFieldsError   = field.error.isDefined && !anySpecificError
+
+      val dayErrorClass   = if (dayError || allFieldsError) errorClass else ""
+      val monthErrorClass = if (monthError || allFieldsError) errorClass else ""
+      val yearErrorClass  = if (yearError || allFieldsError) errorClass else ""
 
       val items = Seq(
         InputItem(
@@ -35,21 +45,21 @@ trait DateFluency {
           name    = s"\${field.name}.day",
           value   = field("day").value,
           label   = Some(messages("date.day")),
-          classes = s"govuk-input--width-2 \$errorClass".trim
+          classes = s"govuk-input--width-2 \$dayErrorClass".trim
         ),
         InputItem(
           id      = s"\${field.id}.month",
           name    = s"\${field.name}.month",
           value   = field("month").value,
           label   = Some(messages("date.month")),
-          classes = s"govuk-input--width-2 \$errorClass".trim
+          classes = s"govuk-input--width-2 \$monthErrorClass".trim
         ),
         InputItem(
           id      = s"\${field.id}.year",
           name    = s"\${field.name}.year",
           value   = field("year").value,
           label   = Some(messages("date.year")),
-          classes = s"govuk-input--width-4 \$errorClass".trim
+          classes = s"govuk-input--width-4 \$yearErrorClass".trim
         )
       )