+++ src/main/g8/project/AppDependencies.scala
@@ -2,12 +2,12 @@ import sbt._
 
 object AppDependencies {
 
-  private val bootstrapVersion = "7.22.0"
-  private val hmrcMongoVersion = "1.2.0"
+  private val bootstrapVersion = "7.23.0"
+  private val hmrcMongoVersion = "1.4.0"
 
   val compile = Seq(
     play.sbt.PlayImport.ws,
-    "uk.gov.hmrc"       %% "play-frontend-hmrc"             % "7.9.0-play-28",
+    "uk.gov.hmrc"       %% "play-frontend-hmrc"             % "7.28.0-play-28",
     "uk.gov.hmrc"       %% "play-conditional-form-mapping"  % "1.13.0-play-28",
     "uk.gov.hmrc"       %% "bootstrap-frontend-play-28"     % bootstrapVersion,
     "uk.gov.hmrc.mongo" %% "hmrc-mongo-play-28"             % hmrcMongoVersion