+++ src/main/scaffolds/datePage/app/controllers/$className$Controller.scala
@@ -26,11 +26,11 @@ class $className$Controller @Inject()(
                                         view: $className$View
                                       )(implicit ec: ExecutionContext) extends FrontendBaseController with I18nSupport {
 
-  def form = formProvider()
-
   def onPageLoad(mode: Mode): Action[AnyContent] = (identify andThen getData andThen requireData) {
     implicit request =>
 
+      val form = formProvider()
+      
       val preparedForm = request.userAnswers.get($className$Page) match {
         case None => form
         case Some(value) => form.fill(value)
@@ -42,6 +42,8 @@ class $className$Controller @Inject()(
   def onSubmit(mode: Mode): Action[AnyContent] = (identify andThen getData andThen requireData).async {
     implicit request =>
 
+      val form = formProvider()
+      
       form.bindFromRequest().fold(
         formWithErrors =>
           Future.successful(BadRequest(view(formWithErrors, mode))),