+++ src/main/scaffolds/datePage/app/forms/$className$FormProvider.scala
@@ -1,14 +1,15 @@
 package forms
 
-import java.time.LocalDate
-
 import forms.mappings.Mappings
-import javax.inject.Inject
 import play.api.data.Form
+import play.api.i18n.Messages
+
+import java.time.LocalDate
+import javax.inject.Inject
 
 class $className$FormProvider @Inject() extends Mappings {
 
-  def apply(): Form[LocalDate] =
+  def apply()(implicit messages: Messages): Form[LocalDate] =
     Form(
       "value" -> localDate(
         invalidKey     = "$className;format="decap"$.error.invalid",